refactor(cli/services/login): split into 9-module package preserving re-export surface
P3.T4 — pure structural split of the 1337-LOC ``cli/services/login.py``
into the ``cli/services/login/`` package along the leaf-ward DAG declared
in ``phase-3.md``. No behaviour change; ``console.print`` calls move
unchanged with their owning functions (Source-Plan Override: rev-2
dropped the "no console.print in service layer" requirement for T4 —
that's a separate I/O refactor and explicitly out of scope here).

Module layout (all submodules ≤ 400 LOC):

  cookie_domains.py     151  leaf — _INCLUDE_DOMAINS_ALL + parse/resolve/build
  rookiepy_errors.py     31  leaf — _handle_rookiepy_error (console only)
  cookie_jar.py         156  → rookiepy_errors  (the helper is unused here)
  chromium_accounts.py  209  → cookie_jar, rookiepy_errors, cookie_domains
  firefox_accounts.py   129  → cookie_jar*, rookiepy_errors, cookie_domains
  browser_accounts.py   234  → chromium_accounts, firefox_accounts,
                              cookie_jar, rookiepy_errors, cookie_domains
  profile_targets.py    136  leaf — _PROFILE_NAME_RE + _validate_profile_name
                              + email_to_profile_name + alloc helpers
  cookie_writes.py      168  _write_extracted_cookies + _select_account
                              + _select_refresh_account
  refresh.py            337  _login_*_*, _refresh_from_browser_cookies,
                              _sync_server_language_to_config
  __init__.py           114  re-export-only (AST-guarded)

(* the cookie_jar edge is allowed by the DAG but the firefox helpers
don't actually call into it — they delegate to firefox-containers and
hand raw cookies back to ``_enumerate_one_jar`` at the dispatcher.)

The DAG diagram in phase-3.md routes ``cookie_domains`` only through
the chromium/firefox subordinates, but ``_read_browser_cookies``'s
"auto" + named-alias branch (the legacy ``rookiepy.load`` path) builds
its own domain list at the dispatcher in ``browser_accounts``. That edge
is added to ``ALLOWED_EDGES`` in the DAG lint with an inline comment;
the graph stays acyclic (``cookie_domains`` is a leaf).

Patch-surface preservation
--------------------------

The package ``__init__.py`` re-exports:

* The 13 baseline names from ``tests/_fixtures/session_reexport_baseline.txt``
  (the ``from .services.login import (…)`` block in ``cli/session_cmd.py``).
  Verified by ``tests/unit/cli/test_session_patch_surface.py``: all 13
  names are importable, callable, and monkey-patchable through both the
  session module's namespace and the package namespace.
* ``_PROFILE_NAME_RE`` + ``_validate_profile_name`` + ``email_to_profile_name``
  — accessed by ``cli/profile_cmd.py`` via ``login_service.<name>``.
* External re-imports (``console``, ``get_storage_path``, ``run_async``,
  ``NotebookLMClient``, ``fetch_tokens_with_domains``) so legacy patch
  targets like ``"notebooklm.cli.services.login.get_storage_path"`` keep
  resolving. **Important caveat documented in the package docstring:**
  patching these on the package no longer reaches the submodule callers
  (each submodule binds its own imports). Tests that need
  call-site interception use the updated ``patch_session_login_dual``
  fixture in ``tests/_fixtures/cli_session.py``, which now fans the
  patch out to every submodule that binds the name.

37+ legacy ``monkeypatch.setattr("notebooklm.cli.session_cmd.<name>", …)``
sites continue to work byte-for-byte — verified by the full unit +
integration test suite (5990 passed, 12 skipped).

Test-side changes (mechanical, no behaviour shift)
--------------------------------------------------

* ``tests/_fixtures/cli_session.py`` — ``patch_session_login_dual`` now
  iterates ``notebooklm.cli.services.login``'s submodules and fans the
  same mock object out to every submodule that binds the requested
  name. The dual-patch API stays the same; the patch surface grew from
  2 modules (session + package) to N (session + package + every submodule
  that imports the name). This restores the pre-split "patch once,
  intercept everywhere" contract.

* ``tests/integration/cli_vcr/test_login_browser_cookies.py`` —
  ``_read_browser_cookies`` is now defined in ``browser_accounts`` and
  called from ``refresh._login_with_browser_cookies``; both binding
  sites need the patch (in addition to the package re-export) so the
  cookie reader can be stubbed out.

* ``tests/unit/cli/test_session_edge_cases.py::test_windows_storage_chmod_skipped``
  — the ``sys.platform != "win32"`` guard now lives in two submodules
  (``cookie_writes`` and ``refresh``); the inspect-source assertion was
  updated to check both.

* ``src/notebooklm/cli/error_handler.py`` — ``ALLOWED_CLICK_EXCEPTION_SITES``
  entry updated from ``services/login.py:54`` to
  ``services/login/profile_targets.py:26``.

* ``CLAUDE.md`` — services/ subtree updated to reflect the package layout
  (required by the ``test_real_claude_md`` freshness lint).

Acceptance
----------

* ``cli/services/login.py`` deleted; ``cli/services/login/`` is a package.
* Every non-``__init__.py`` file ≤ 400 LOC.
* ``__init__.py`` passes the AST re-export-only guard.
* DAG lint passes: 9 implementation modules + the strict leaf-ward
  edge set; cycle detector returns ``[]``.
* ``python -c "import notebooklm.cli.services.login"`` succeeds.
* ``python -m pytest tests/unit/cli/test_login.py -x`` — 47 passed.
* Full unit + integration suite — 5990 passed, 12 skipped.
* ``uv run ruff check src/notebooklm/cli`` — clean.
* ``uv run mypy src/notebooklm/cli --ignore-missing-imports`` — clean.
